# backend/api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/v1/chats/{chatId}/read:
    post:
      tags:
      - api-chat-controller
      summary: Mark messages in a chat as read
      operationId: markMessagesAsRead
      parameters:
      - name: chatId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Chat not found
        "200":
          description: Messages marked as read
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - User not part of this chat
...
  /api/v1/chats:
    get:
      tags:
      - api-chat-controller
      summary: Get user's chats
      operationId: getUserChats
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        "200":
          description: Chats found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat_ChatListBasicView"
        "401":
          description: Unauthorized
        "404":
          description: No chats found
  /api/v1/chats/{chatId}/messages:
    get:
      tags:
      - api-chat-controller
      summary: Get messages from a specific chat
      operationId: getChatMessages
      parameters:
      - name: chatId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Messages found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message_ChatMessageView"
        "404":
          description: Chat not found
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - User not part of this chat
  /api/v1/chats/unread-count:
    get:
      tags:
      - api-chat-controller
      summary: Get count of unread messages
      operationId: getUnreadCount
      responses:
        "200":
          description: Unread count retrieved
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "401":
          description: Unauthorized
...
  /api/v1/chats/{chatId}:
    delete:
      tags:
      - api-chat-controller
      summary: Delete a chat
      operationId: deleteChat
      parameters:
      - name: chatId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Chat not found
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - User not part of this chat
        "200":
          description: Chat deleted successfully
components:
  schemas:
...
    Chat_ChatListBasicView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        user1:
          $ref: "#/components/schemas/User_ChatListBasicView"
        user2:
          $ref: "#/components/schemas/User_ChatListBasicView"
        messages:
          type: array
          items:
            $ref: "#/components/schemas/Message_ChatListBasicView"
        lastMessage:
          $ref: "#/components/schemas/Message_ChatListBasicView"
        lastMessageTime:
          type: string
          format: date-time
        unreadCount:
          type: integer
          format: int64
    Message_ChatListBasicView:
      type: object
      properties:
        sender:
          $ref: "#/components/schemas/User_ChatListBasicView"
        content:
          type: string
    Post_ChatListBasicView:
      type: object
    Reply_ChatListBasicView:
      type: object
    User_ChatListBasicView:
      type: object
      properties:
        username:
          type: string
        upvotedPosts:
          type: array
          items:
            $ref: "#/components/schemas/Post_ChatListBasicView"
        downvotedPosts:
          type: array
          items:
            $ref: "#/components/schemas/Post_ChatListBasicView"
        likedReplies:
          type: array
          items:
            $ref: "#/components/schemas/Reply_ChatListBasicView"
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post_ChatListBasicView"
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Reply_ChatListBasicView"
    Message_ChatMessageView:
      type: object
      properties:
        id:
          type: integer
          format: int64
        sender:
          $ref: "#/components/schemas/User_ChatMessageView"
        receiver:
          $ref: "#/components/schemas/User_ChatMessageView"
        content:
          type: string
        read:
          type: boolean
        timestamp:
          type: string
          format: date-time
    Post_ChatMessageView:
      type: object
    Reply_ChatMessageView:
      type: object
    User_ChatMessageView:
      type: object
      properties:
        username:
          type: string
        upvotedPosts:
          type: array
          items:
            $ref: "#/components/schemas/Post_ChatMessageView"
        downvotedPosts:
          type: array
          items:
            $ref: "#/components/schemas/Post_ChatMessageView"
        likedReplies:
          type: array
          items:
            $ref: "#/components/schemas/Reply_ChatMessageView"
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post_ChatMessageView"
        replies:
          type: array
          items:
            $ref: "#/components/schemas/Reply_ChatMessageView"
...